+++ js/firebaseConfig.js
@@ -1,0 +1,10 @@
+export let firebaseConfig = {
+  apiKey: "AIzaSyC7uD176kMJTedZtbl4EDtbs2hZoDT057M",
+  authDomain: "housepurchase.firebaseapp.com",
+  projectId: "housepurchase",
+  storageBucket: "housepurchase.appspot.com",
+  messagingSenderId: "40016422170",
+  appId: "1:40016422170:web:6fae180d6f615f49a138a3",
+};
+// Initialize Firebase
+firebase.initializeApp(firebaseConfig);
