+++ sass/_header.scss
@@ -1,7 +1,7 @@
 .header {
 
     height: 10vh;
-    width: 100vw;
+    width: 100%;
     background-size: cover;
     margin: 0;
     padding: 0;
