+++ sass/components/_form.scss
@@ -1,14 +1,24 @@
+.container-inner {
+    height: 100vh;
+    display: grid;
+    justify-items: center;
+    grid-template-columns: 1fr 1fr 1fr 1fr;
+    grid-template-rows: 1rem 6rem 1fr;
+    grid-row-gap: 1rem;
+    grid-column-gap: 1.2rem;
+    
+}
+
+
 .firstForm {
-    display: inline-block;
-    width: auto;
-    margin: 4rem 4rem;
-   
+    grid-row: 2 / 3;
+    width: 27rem;
+    margin: 0.1rem 0.5rem;
     position: relative;
     animation: moveUp 2s ease-out;
     display: none;
-
     &__container {
-  
+
         background-color: rgba($color-grey-dark-2, 0.4);
         border-radius: 1rem;
         padding: 1rem 0;
@@ -38,6 +48,13 @@
     }
 
    }
+
+.call-modal { 
+    grid-row: 2/3;
+    // grid-column: 2 / 3;
+    @include button;
+    justify-self: center;
+}
