+++ sass/components/_modal.scss
@@ -1,4 +1,4 @@
-.houseModal {
+.house-modal {
     height: 100vh;
     width: 100%;
     position: fixed;
@@ -6,7 +6,7 @@
     left: 0%;
     background-color: rgba($color-black, .8); // make sure no space after rgba
     z-index: 999999;
-   
+    display: none;
 
     &__content {
         @include centerItem;
